+++ previous-versions/sync-for-expenses-version-1/pkg/models/shared/synccompletewebhook.go
@@ -13,9 +13,11 @@ type SyncCompleteWebhook struct {
 	// Unique identifier for your SMB in Codat.
 	CompanyID *string                  `json:"CompanyId,omitempty"`
 	Data      *SyncCompleteWebhookData `json:"Data,omitempty"`
-	// A human readable message about the webhook.
+	// A human-readable message about the webhook.
 	Message *string `json:"Message,omitempty"`
 	// Unique identifier for the rule.
+	//
+	// Deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
 	RuleID *string `json:"RuleId,omitempty"`
 	// The type of rule.
 	RuleType *string `json:"RuleType,omitempty"`