+++ previous-versions/sync-for-expenses-version-1/pkg/models/shared/syncstartedwebhook.go
@@ -3,64 +3,10 @@
 package shared
 
 type Data struct {
-	// In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
-	//
-	// ```
-	// 2020-10-08T22:40:50Z
-	// 2021-01-01T00:00:00
-	// ```
-	//
-	//
-	//
-	// When syncing data that contains `DateTime` fields from Codat, make sure you support the following cases when reading time information:
-	//
-	// - Coordinated Universal Time (UTC): `2021-11-15T06:00:00Z`
-	// - Unqualified local time: `2021-11-15T01:00:00`
-	// - UTC time offsets: `2021-11-15T01:00:00-05:00`
-	//
-	// > Time zones
-	// >
-	// > Not all dates from Codat will contain information about time zones.
-	// > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
-	SyncDateRangeFinishUtc *string `json:"SyncDateRangeFinishUtc,omitempty"`
-	// In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
-	//
-	// ```
-	// 2020-10-08T22:40:50Z
-	// 2021-01-01T00:00:00
-	// ```
-	//
-	//
-	//
-	// When syncing data that contains `DateTime` fields from Codat, make sure you support the following cases when reading time information:
-	//
-	// - Coordinated Universal Time (UTC): `2021-11-15T06:00:00Z`
-	// - Unqualified local time: `2021-11-15T01:00:00`
-	// - UTC time offsets: `2021-11-15T01:00:00-05:00`
-	//
-	// > Time zones
-	// >
-	// > Not all dates from Codat will contain information about time zones.
-	// > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
-	SyncDateRangeStartUtc *string `json:"SyncDateRangeStartUtc,omitempty"`
 	// Unique identifier for the failed sync.
 	SyncID *string `json:"syncId,omitempty"`
 	// The type of sync being performed.
 	SyncType *string `json:"syncType,omitempty"`
-}
-
-func (o *Data) GetSyncDateRangeFinishUtc() *string {
-	if o == nil {
-		return nil
-	}
-	return o.SyncDateRangeFinishUtc
-}
-
-func (o *Data) GetSyncDateRangeStartUtc() *string {
-	if o == nil {
-		return nil
-	}
-	return o.SyncDateRangeStartUtc
 }
 
 func (o *Data) GetSyncID() *string {
@@ -88,9 +34,11 @@ type SyncStartedWebhook struct {
 	// Unique identifier for your SMB in Codat.
 	CompanyID *string `json:"CompanyId,omitempty"`
 	Data      *Data   `json:"Data,omitempty"`
-	// A human readable message about the webhook.
+	// A human-readable message about the webhook.
 	Message *string `json:"Message,omitempty"`
 	// Unique identifier for the rule.
+	//
+	// Deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
 	RuleID *string `json:"RuleId,omitempty"`
 	// The type of rule.
 	RuleType *string `json:"RuleType,omitempty"`